fixed repeated character bug
If a word had multiple of the same character it would print out that character the number of times it appeared in the word if the user guessed the letter in the wrong place. This was due to a for loop not being broken out of after determining if the letter was in the word. The for loop essentially evaluated to true the number of times the letter appeared in the word and consequently printed the character multiple times.